Refactor the GUI hex view and pane components (#427)
---------
