+++ frontend/src/hex/stores.js
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export const screenHeight = writable(0);
