+++ frontend/src/ofrak/remote_resource.js
@@ -170,10 +170,9 @@ export class RemoteResource extends Resource {
     if (this.data_id === null) {
       return null;
     }
-    let result = await fetch(`${this.uri}/get_data_length`).then((r) =>
-      r.json()
-    );
-    return result;
+    return await fetch(`${this.uri}/get_data_length`).then(async (r) => {
+      return await r.json();
+    });
   }
 
   async get_data_range_within_parent() {
