+++ frontend/src/stores.js
@@ -49,3 +49,5 @@ export function loadSettings(forceReset) {
 }
 
 export let settings = writable(loadSettings());
+
+export let resourceNodeDataMap = writable({});
