+++ frontend/src/utils/SearchBar.svelte
@@ -41,6 +41,7 @@
     width: 100%;
     padding-bottom: 1em;
     min-height: 2.25em;
+    background-color: var(--main-bg-color);
   }
 
   .resultwidgets {
@@ -93,6 +94,9 @@
     prevOptions = {};
 
   function nextMatch() {
+    if (searchResults.matches === undefined) {
+      return searchResults;
+    }
     let nextIndex = searchResults.index + 1;
     if (nextIndex >= searchResults.matches.length) {
       nextIndex = 0;
@@ -113,7 +117,8 @@
       searchQuery == prevQuery &&
       searchOptions.searchType == prevOptions.searchType &&
       searchOptions.regex == prevOptions.regex &&
-      searchOptions.caseIgnore == prevOptions.caseIgnore
+      searchOptions.caseIgnore == prevOptions.caseIgnore &&
+      searchResults.matches != undefined // Search results are the only attribute of the search we have control of from other components, so we use it to clear the search when changing selected resources.
     );
   }
 
@@ -163,6 +168,7 @@
       </option>
     {/each}
   </select>
+  <!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
   <form
     on:submit="{async (e) => {
       e.preventDefault();
