+++ frontend/src/views/AssemblyView.svelte
@@ -12,6 +12,9 @@
     flex-wrap: nowrap;
     justify-content: flex-start;
     align-items: stretch;
+    min-height: calc(100% - 6em);
+    max-height: calc(100% - 6em);
+    overflow: auto;
   }
 
   .horizontal-spacer {
