+++ frontend/src/views/AttributesView.svelte
@@ -3,6 +3,12 @@
     font-size: 1em;
   }
 
+  .attributes {
+    max-height: calc(100% - 3em);
+    min-height: calc(100% - 3em);
+    overflow: auto;
+  }
+
   a {
     color: var(--main-fg-color);
   }
@@ -35,19 +41,20 @@
   }
 </script>
 
-{#if resource !== undefined}
-  <h1>Tags:</h1>
-  {#each resource.get_tags() as tag, i}
-    <a
-      href="{'https://ofrak.com/docs/reference/' +
-        tag.split('.').slice(0, -1).join('/') +
-        '.html#' +
-        tag}"
-      target="_blank"
-      rel="noreferrer">{cleanOfrakType(tag)}</a
-    >{i !== resource.get_tags().length - 1 ? ", " : ""}
-  {/each}
-
-  <h1>Attributes:</h1>
-  <StructuredList object="{attributes}" />
-{/if}
+<div class="attributes">
+  {#if resource !== undefined}
+    <h1>Tags:</h1>
+    {#each resource.get_tags() as tag, i}
+      <a
+        href="{'https://ofrak.com/docs/reference/' +
+          tag.split('.').slice(0, -1).join('/') +
+          '.html#' +
+          tag}"
+        target="_blank"
+        rel="noreferrer">{cleanOfrakType(tag)}</a
+      >{i !== resource.get_tags().length - 1 ? ", " : ""}
+    {/each}
+    <h1>Attributes:</h1>
+    <StructuredList object="{attributes}" />
+  {/if}
+</div>
